Commiting changes from timer
