+++ App.js
@@ -8,7 +8,7 @@ import GenerateLettersComponent from "./components/GenerateLettersComponent"
 import TimerComponent from "./components/TimerComponent"
 import ScoresComponent from "./components/ScoresComponent"
 import DebugComponent from "./components/DebugComponent"
-import Timer2Component from "./components/Timer2Component"
+import CountdownComponent from "./components/CountdownComponent"
 
 
 const Bamboozle = StackNavigator({
@@ -27,6 +27,9 @@ const Bamboozle = StackNavigator({
   Debug: {
     screen: DebugComponent
   },
+  Countdown: {
+    screen: CountdownComponent
+  }
 
 }, {
   headerMode: 'screen'
